add inpaint pipe
